requirements.txt
fastapi
mysql-connector-python
openai
python-dotenv
python-multipart
requests
scipy
sounddevice
uvicorn
